app/frontend/src/app/game/components/2_navigasi_menu/2_navigasi_bawah/5_geopolitik/1_PBB/1_pemungutan_suara/0_pemungutan_AI_logic/requirements.txt
python>=3.8
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.0
